requirements.txt
# --- 核心运行依赖 ---
# 异步网络请求 (核心引擎)
httpx>=0.28.0

# API服务器框架 (server/api.py)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# HTML/XML解析 (processors/html_worker.py, epub_worker.py)
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 基础工具
python-dotenv>=1.0.0            # 读取 .env 配置
aiofiles>=23.0.0                # 异步文件操作
tqdm>=4.66.0                    # 进度条 (保留着以防万一)

# --- 开发工具 (可选，保留方便调试) ---
# 只有在运行测试或格式化代码时才需要
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0